chore(gh): Tweak Bug template
This is a follow-up on
https://github.com/kirillbobyrev/pabi/commit/3c9f332362cfe381911e4937a6b7aee9ff225b24

The Bug template was introduced and contains the most important fields &
provides a great baseline. This is a minor tweak to shorten the template
(to reduce the burden of typing all the info) and specify an OS (if
possible). Of course, this is still an unused template. When (and
hopefully if) the engine is used and the bugs are reported, there will
be more context and understanding of what else could be added.