feat(movegen): Implement Move Generation MVP
It seems to be working in the basic cases. The code is very dense and
needs to be improved in the future but it's a start I can base of. The
performance is not known and the test infrastructure is yet to come: the
next step is implementing move application and then checking against
existing Perft results.